a11y: make each page have a unique `<title>`
- each page have a unique title suffixed with the site name. Example:
"Votre mot de passe - MonComptePro"
- if a 404 error or other http error is detected, the page has the error
info in the title ("Erreur 404 - MonComptePro")
- if an error notification is set, prefix the `<title>` with "Erreur"

This makes navigation with screen readers better, as screen reader users
often rely on the document's page as a first hint to what page they are
on. Checking for notifications to mark pages as "Erreur" is also useful
to let them know directly if there is a problem. When filling a form,
posting it, and it fails, the first feedback they have is now that there
is an error.

It's also beneficial for everyone as it's easier to notice at a glance
where we are by just looking at the tab without having to focus the
page.

RGAA criteria 8.5 and 8.6.