fix: make the site layout consistent accross all pages:
using grid layout for the `<body>` made the layout move between pages
because the main content never has the same height. Now we have a
simpler layout, just telling the footer to stay at the bottom, header
and main always start at the same place.

this is also useful in case we want to add global stuff in the body
later without breaking everything. Skip links will be added and will
benefit from this simpler layout, as they are shown/hidden based on
navigation.