a11y: separate each region for better screen readers support
screen reader have shortcuts to go from region to region (header, nav,
main, footer and elements with role="region").

Here we make sure each site region is not a child of another.

Because:

- some buggy screen reader versions dont list regions that are inside
other regions. I admit this is a rare case but if it's easy to handle
this case, why not, it's better to fix it
- it makes more sense semantically to have the site `<nav>` outside
the `<main>`
- it makes more sense semantically to have the site heading
"MonComptePro" inside the `<header>`
- now we can stop copy/pasting the "MonComptePro" heading in every new
page that uses the same layout

Visuals don't change thanks to a couple CSS changes. Only semantics do.

For pages not using the dashboard layout, nothing is changed as markup
was already good.

RGAA criteria 9.2