a11y: fix headings hierarchy
- remove the gaps between headings. Gaps are when you have a section
starting with a h3, and a direct part of this section starts with a h5,
instead of a h4. Lots of screen reader users rely on the headings
structure to understand page content. Depending on content, having gaps
in the hierachy can make the content confusing to understand. Here the
problem is not that impactful as content is pretty simple, but it's
better to fix anyway!
- start every page at h1. I'd admit this is not an enormous issue but it
makes more semantic sense to start content at h1 rather than h4. Plus,
for future devs, it allows for more content structure (we can go deeper
in the hierarchy if starting at h1 than h4).

RGAA criteria 9.1