a11y: make external links explicit to screen reader users
currently, links opening new windows tell the user they do with an icon.
But screen reader users don't have the info.

this tiny script automatically labels links to tell screen reader users
just that.

I chose to use a script rather than editing every link so that we don't
have to think about it all the time. We could also go the "manual" route
and fix every html anchor ourselves without JS.

there is no explicit RGAA criteria that says that everyone should know
that an anchor will open a tab. But if we decide to show that to sighted
users, we should definitely not prevent screen reader users to get the
same information.