a11y: add context to interactive elements for screen reader users
One good thing to do for screen reader users is to make every
interactive element understandable on its own. Because people can
navigate from anchor to anchor, or from button to button. It's a way to
navigate quickly accross the page or get a glimpse of what is on a page.

So, in a list setting, if 4 buttons have a "Delete" label, you dont
really know what deletes what without having to navigate around the
button.

This adds context in each necessary label.

RGAA criteria 6.1, 11.2, 11.9.