a11y: remove unwanted `<br>` that are annoying for screen reader users
this cleans up usage a `<br>` tags that were used only for visual
purposes and that were a little nuisance with a screen reader.

NVDA for example, vocalizes `<br>` as "Blank". More often that not, we
don't want that. We mostly want to mark specific paragraphs.

So this does a few things:

- remove brs that didn't do anything visually (there were a few, I guess
some old code?)
- use dsfr spacing classes instead of brs that were used only for
presentation
- use `<p>` instead of `<div>` + `<br>` for some blocks of texts. Some
screen readers will better vocalize text with paragraph tags.
- hide from screen readers the brs that are necessary to be in the DOM
but that we don't want to vocalize. For example:

```
République
-<br>Française
+<br aria-hidden="true">Française
```

is done because, without hiding the brs here, the screen reader will
pause. It'll say "République." Then we have to interact to ask for the
next text. Then "Française".
Now it correctly says "République Française" without pausing, and
without breaking the visual need of the br.

RGAA criteria 8.9