a11y: be more explicit on the footer home link that it logouts
The logo in the footer is a link that disconnects the user. Be explicit
about that in the title (for mouse users) and aria-label (for screen
reader users).

We might want to be event more explicit, because for now mobile/keyboard
users dont get the info.

Or maybe this anchor could link to home, instead of logout? Because
logout is already there in the header.