a11y: in notifications, be more explicit on errors/warnings
the nature of the notification is only shown via color/icon. We could
argue it's sufficient for sighted users.. we could also argue otherwise,
because sometimes icons are not fully understood, and we can't rely on
color to explain something on its own (if we consider the icon is not
enough to understand).

Screen reader users don't get the info at all about the what the message
is about. We could add an "Error" or "Warning" prefix only for screen
reader users.

Here we decide to follow the DSFR instructions and be 100% explicit
about the nature of the message by showing it for everyone.

This follows the DSFR instructions for errors, warnings and info. I'd
argue the "success" type is a bit weird to prefix with "Succès"? Leaving
it as is.