a11y: be more explicit for screen reader users with autosubmit forms
Autosubmitting forms show a loading spinner but it doesn't have a
text alternative. For now we rely exclusively on the actionLabel
passed to the component.

This action label can be whatever we want depending on use case. Maybe
sometimes the label will let correctly know the user that somethings is
loading, sometimes we could argue it's not 100% explicit.

Adding a text alternative for the loading spinner fixes that.

RGAA criteria 1.1