a11y: make main nav use default screen reader label
Saying "Menu latéral" is not that useful for screen reader users as they
don't care about the position of the menu.

I'd argue the best thing to do here is just to not specify a label.
Screen reader will vocalize the default navigation label that is the one
often used for main navigations.