a11y: have same form labels for sighted and screen reader users
there were a few tiny things about form labels:

- some label/input connections, through input[id] and label[for], were
not correct
- a few input labels were different for screen reader users. Sometimes
there is no real benefit to doing that, because we miss the user help of
label hint texts.

For required inputs marked with "*" visually, it's still cool to have a
specific aria-label that doesn't have it, because screen reader users
don't rely on it, they rely on [required] attributes.