a11y: tell users what the "*" means in forms with some required fields
the "*" is not necessarily understood implicitly by everyone. We must
explicitly tell the user about what it means, before showing the form.

Note we hide this from screen reader users, as the screen reader takes
care of saying what is required thanks to `required` attributes. Plus,
we dont vocalize "*" for screen readers, it would be confusing.

RGAA criteria 11.10