refactor: a11y: make password help texts better for screen reader users
we have a few rules on how to write a password. The idea is to help
screen reader users as they type. But we should care about not "over
helping", we don't want to spam too much.

The idea is:

- the password input is tied to whole block explaining everything via
aria-describedby. Screen reader users get the whole stuff in
aria-describedby, whatever the state of the input.
- as the user types, an alert warns the user about its progress on
making a good password. Something like "conditions remaining: 1 number,
1 uppercase" etc. We debounce the alert in order to prevent spamming.
This is done via an aria-live zone used exclusively for alerts.

I allowed myself to refactor a bit the code, hoping to make it easier to
work with.

Note: the commas inside "fr-sr-only" are there so that the
aria-describedby announcements pause between each item. Because, if the
user navigates to the list, each item is vocalized one by one. But if
the user focuses the input, the aria description is stripped of its
semantics. So we use commas so that the screen reader understands it
has to pause.