feat: retrieve legal content pages at runtime instead of buildtime
change the logic of previous commit: now we get fresh content of
external pages when we request it in the controller and cache it
for a day in the DB.

not sure of what the best logic is. This better fits current workflow
at the cost of a tiny bit slower request once a day to rebuild the
cache.
Previous commit ensured quick loading all the time but required to
rebuild the content regularly somehow.