feat: following up PR#500, changing how we handle SR form labels
In form with required fields with "*" showing that to users, we doubled
labels: a visible one in <label>, containing the "*" character, and a
`aria-label` that was the exact same one, without the "*" character.

That was to prevent screen reader users to hear the asterisk, while they
dont need to: the screen reader says when a field is required.

But I noticed that mistakes were easy to make with that way of doing
this. We can easily change a visible label and forget to change the SR
one.

Now there is no specific screen reader label, we use aria-hidden around
the asterisk to handle correct screen reader support.