a11y: do not make screen reader users think org suggestions are links
so this "fix" is not 100% ideal but I'd say an acceptable way of
improving things.

The problem was: we use DSFR cards, that are meant to be used with <a>
tags. But our cards dont target other pages: the <a> in them actually
trigger something on the page, like a button. So they have empty hrefs,
<a> is used mostly for the DSFR styling.

This is an issue because with a screen reader, it's announced that those
are links, so I expect to click on them and be redirected to a new page.

We can't use actual <button> tags because cards are not really meant for
that in DSFR and if we do, there are styling issues.

So the trick is to put a role "button" on the <a> tag so that it's
announced as a button for screen readers. And try to match a bit
the behavior of a button, by adding the space key trigger possible.

It's not ideal because we do not match 100% the behavior of a button.
But it's acceptable because I'd say the missing pieces are not a
deal-breaker.